requirements.txt
python-dotenv==1.0.0
langchain==0.0.325
openai==0.28.1
chromadb===0.4.15
tiktoken===0.5.1